requirements.txt
py-clob-client
python-dotenv
requests
web3
flask
gunicorn
